Add confirmation with SweetAlert2 to delete bills and projects

+++ src/app/bill-card/bill-card.component.ts
@@ -2,6 +2,8 @@ import { Component, Input, Output, EventEmitter } from '@angular/core'; // Impor
 import { CommonModule } from '@angular/common';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { FirebaseService } from '../firebase.service';
+import Swal from 'sweetalert2';
+
 
 @Component({
   selector: 'app-bill-card',
@@ -21,16 +23,37 @@ export class BillCardComponent {
   }
 
   deleteBill() {
-    if (this.bill && this.bill.id) {
-      this.firebaseService.deleteBill(this.bill.id)
-        .then(() => {
-          console.log('Bill deleted successfully');
-          
-          // Emit the event to notify the parent component that the bill was deleted
-          this.billDeleted.emit(this.bill.id);
-        })
-        .catch((error) => {
-          console.error('Error deleting bill:', error);
-        });
-    }}
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "¡Esta acción no se puede deshacer!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#56a6e8",
+      cancelButtonColor: "#e85d56",
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar",
+    }).then((result) => {
+      if (result.isConfirmed && this.bill && this.bill.id) {
+        this.firebaseService.deleteBill(this.bill.id)
+          .then(() => {
+            Swal.fire({
+              title: "¡Eliminado!",
+              text: "La factura ha sido eliminada.",
+              icon: "success",
+            });
+            // Emit the event to notify the parent component that the bill was deleted
+            this.billDeleted.emit(this.bill.id);
+          })
+          .catch((error) => {
+            Swal.fire({
+              title: "Error",
+              text: "Hubo un problema al eliminar la factura.",
+              icon: "error",
+            });
+            console.error('Error deleting bill:', error);
+          });
+      }
+    });
+  }
+  
 }
